Set CircleCI to build with Node.js 8 and 9
The idea is to build with all supported major Node.js versions.

I chose to duplicate the job config for each Node.js version completely
instead of deriving both configs from a common base. The latter was hard
to do given that I wanted to have a different package cache key in each
config (so that packages installed using Node.js 8 are not used for
Node.js 9 builds and vice versa).