.github/workflows: Replace `npm install` by `npm ci` where possible
The `npm ci` command is recommended for installing dependencies in CI
environment for a variety of reasons (see its documentation).

The command deletes `node_modules` if present to ensure a clean install,
so its cache is removed.

In Node.js 9, `npm install` is still used because of old `npm` version
that doesn't support `npm ci`.